Move GPIO into camera handler :hammer:
